rl: store rollout frames as float16 to halve CPU RAM usage
Each rollout buffer holds 200 × (3 frames × 3×224×224) float32 tensors
≈ 360 MB.  With 16 rollouts that is ~5.8 GB of frame data, which
combined with the CLIP + VERA weights (~1.3 GB) and Python/CUDA overhead
saturated the Colab 12 GB RAM limit and triggered SIGKILL at rollout 15.

Fix: store frames as torch.float16 in buf.add() and restore float32
when moving each mini-batch to GPU in rl_update_batch().  The precision
loss is negligible (CLIP-normalised values sit in [-2, 2], well within
float16's representable range).  Peak RAM for 16 rollouts drops from
~7.1 GB to ~4.2 GB, giving ~3 GB of headroom on a standard Colab T4.

+++ training/rl_trainer_vera.py
@@ -210,7 +210,7 @@ def collect_rollout(
             )
 
         buf.add(
-            frame        = frames_t.cpu(),
+            frame        = frames_t.half().cpu(),   # float16: halves RAM (5.8→2.9 GB/16 rollouts)
             lang_tok     = lang_tok,
             act_hist     = act_hist_in.squeeze(0).cpu(),
             rew_hist     = rew_hist_in.squeeze(0).cpu(),
@@ -495,7 +495,7 @@ def rl_update_batch(
 
         # Move this mini-batch to GPU
         mb_ret      = returns_cpu     [mb_s:mb_e].to(device)
-        mb_frames   = frames_cpu      [mb_s:mb_e].to(device)
+        mb_frames   = frames_cpu      [mb_s:mb_e].float().to(device)   # restore float32 from stored float16
         mb_lang     = lang_tokens_cpu [mb_s:mb_e].to(device)
         mb_ah       = act_hist_cpu    [mb_s:mb_e].to(device)
         mb_rh       = rew_hist_cpu    [mb_s:mb_e].to(device)
